requirements.txt
bcrypt==4.1.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
google-auth==2.25.2
idna==3.6
kubernetes==28.1.0
oauthlib==3.2.2
openshift-client==1.0.19
paramiko==3.3.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-string-utils==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.18
websocket-client==1.7.0